@@ -17,8 +17,8 @@ platforms:
   linux:
 
 environment:
-  sdk: ">=3.0.1 <4.0.0"
-  flutter: ">=3.7.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
