@@ -28,13 +28,13 @@ class ProgressExampleState extends State<ProgressExample> {
               SizedBox(
                 height: 20,
               ),
-              Wrapper(stepsCompleted: 0, type: ZetaBarType.standard, isThin: false, stateChangeable: true),
+              Wrapper(stepsCompleted: 0, type: ZetaProgressBarType.standard, isThin: false, stateChangeable: true),
               SizedBox(
                 height: 20,
               ),
               Wrapper(
                 stepsCompleted: 0,
-                type: ZetaBarType.indeterminate,
+                type: ZetaProgressBarType.indeterminate,
                 isThin: false,
                 label: "UPLOADING ...",
               ),
@@ -50,7 +50,7 @@ class Wrapper extends StatefulWidget {
   const Wrapper(
       {super.key,
       required this.stepsCompleted,
-      this.type = ZetaBarType.standard,
+      this.type = ZetaProgressBarType.standard,
       this.isThin = false,
       this.rounded = true,
       this.stateChangeable = false,
@@ -58,7 +58,7 @@ class Wrapper extends StatefulWidget {
 
   final int stepsCompleted;
   final bool rounded;
-  final ZetaBarType type;
+  final ZetaProgressBarType type;
   final bool isThin;
   final String? label;
   final bool stateChangeable;
@@ -70,7 +70,7 @@ class Wrapper extends StatefulWidget {
 class _WrapperState extends State<Wrapper> {
   late int stepsCompleted;
   late double progress;
-  late ZetaBarType type;
+  late ZetaProgressBarType type;
 
   @override
   void initState() {
@@ -90,7 +90,7 @@ class _WrapperState extends State<Wrapper> {
 
   void setLoading() {
     setState(() {
-      type = type == ZetaBarType.buffering ? ZetaBarType.standard : ZetaBarType.buffering;
+      type = type == ZetaProgressBarType.buffering ? ZetaProgressBarType.standard : ZetaProgressBarType.buffering;
     });
   }
 
@@ -108,7 +108,7 @@ class _WrapperState extends State<Wrapper> {
         Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            widget.type != ZetaBarType.indeterminate
+            widget.type != ZetaProgressBarType.indeterminate
                 ? FilledButton(onPressed: increasePercentage, child: Text("Increase"))
                 : Container(),
             const SizedBox(width: 40),
