@@ -3,69 +3,37 @@ import 'package:widgetbook/widgetbook.dart';
 
 import 'package:zeta_flutter/zeta_flutter.dart';
 
-WidgetbookComponent textWidgetBook() {
-  final Map<String, TextStyle> allTypes = {
-    'Display large': ZetaTextStyles.displayLarge,
-    'Display medium': ZetaTextStyles.displayMedium,
-    'Display small': ZetaTextStyles.displaySmall,
-    'Heading 1': ZetaTextStyles.heading1,
-    'Heading 2': ZetaTextStyles.heading2,
-    'Heading 3': ZetaTextStyles.heading3,
-    'Title large': ZetaTextStyles.titleLarge,
-    'Title medium': ZetaTextStyles.titleMedium,
-    'Title small': ZetaTextStyles.titleSmall,
-    'Body large': ZetaTextStyles.bodyLarge,
-    'Body medium': ZetaTextStyles.bodyMedium,
-    'Body small': ZetaTextStyles.bodySmall,
-    'Label large': ZetaTextStyles.labelLarge,
-    'Label medium': ZetaTextStyles.labelMedium,
-    'Label small': ZetaTextStyles.labelSmall,
-    'Label indicator': ZetaTextStyles.labelIndicator,
-    'Label tiny': ZetaTextStyles.labelTiny,
-  };
-  return WidgetbookComponent(
-    name: 'Typography',
-    useCases: [
-      WidgetbookUseCase(
-        name: 'Individual text styles',
-        builder: (context) => Text(
-          context.knobs.string(label: 'Text', initialValue: 'The quick brown fox jumps over the lazy dog.'),
-          style: context.knobs
-              .list(
-                label: 'Sizes',
-                labelBuilder: (p0) => allTypes.entries.firstWhere((element) => element.value == p0).key,
-                options: allTypes.values.toList(),
-              )
-              .apply(
-                color: Zeta.of(context).colors.textDefault,
-                fontStyle: FontStyle.normal,
-                decoration: TextDecoration.none,
-              ),
-        ),
-      ),
-      WidgetbookUseCase(
-          name: 'All text styles',
-          builder: (context) => SingleChildScrollView(
-                scrollDirection: Axis.vertical,
-                child: Row(
-                  children: [
-                    Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: allTypes.keys.map(
-                          (e) {
-                            return Text(
-                              e,
-                              style: allTypes[e]?.apply(
-                                color: Zeta.of(context).colors.textDefault,
-                                fontStyle: FontStyle.normal,
-                                decoration: TextDecoration.none,
-                              ),
-                            ).paddingAll(ZetaSpacing.b);
-                          },
-                        ).toList()),
-                  ],
-                ),
-              )),
-    ],
-  );
-}
+const Map<String, TextStyle> allTypes = {
+  'Display large': ZetaTextStyles.displayLarge,
+  'Display medium': ZetaTextStyles.displayMedium,
+  'Display small': ZetaTextStyles.displaySmall,
+  'Heading 1': ZetaTextStyles.heading1,
+  'Heading 2': ZetaTextStyles.heading2,
+  'Heading 3': ZetaTextStyles.heading3,
+  'Title large': ZetaTextStyles.titleLarge,
+  'Title medium': ZetaTextStyles.titleMedium,
+  'Title small': ZetaTextStyles.titleSmall,
+  'Body large': ZetaTextStyles.bodyLarge,
+  'Body medium': ZetaTextStyles.bodyMedium,
+  'Body small': ZetaTextStyles.bodySmall,
+  'Label large': ZetaTextStyles.labelLarge,
+  'Label medium': ZetaTextStyles.labelMedium,
+  'Label small': ZetaTextStyles.labelSmall,
+  'Label indicator': ZetaTextStyles.labelIndicator,
+  'Label tiny': ZetaTextStyles.labelTiny,
+};
+
+Widget typographyUseCase(BuildContext context) => Text(
+      context.knobs.string(label: 'Text', initialValue: 'The quick brown fox jumps over the lazy dog.'),
+      style: context.knobs
+          .list(
+            label: 'Sizes',
+            labelBuilder: (p0) => allTypes.entries.firstWhere((element) => element.value == p0).key,
+            options: allTypes.values.toList(),
+          )
+          .apply(
+            color: Zeta.of(context).colors.textDefault,
+            fontStyle: FontStyle.normal,
+            decoration: TextDecoration.none,
+          ),
+    );
